requirements.txt
invoke==0.10.1
watchdog==0.8.3
pep8==1.6.2
pyflakes==0.8.1

Django==1.8
pytz==2015.4
psycopg2==2.6

django-autoslug==1.7.2
python-slugify==0.1.0

django-bootstrap3==5.2.0
django-datetime-widget==0.9.3
Markdown==2.6.2
djrill==1.3.0
tweepy==3.3.0

celery==3.1.18
celery[redis]
django-redis-cache==0.13.1
